+++ examples/simple.rs
@@ -18,9 +18,9 @@ extern crate pushrod;
 use piston_window::*;
 use pushrod::core::main::*;
 use pushrod::widget::box_widget::*;
+use pushrod::widget::image_widget::*;
 use pushrod::widget::text_widget::*;
 use pushrod::widget::timer_widget::*;
-use pushrod::widget::image_widget::*;
 use pushrod::widget::widget::*;
 
 fn main() {
@@ -32,19 +32,19 @@ fn main() {
     let factory: GfxFactory = window.factory.clone();
     let mut prod: Pushrod = Pushrod::new(window);
 
-//    let mut text_widget = TextWidget::new(
-//        factory,
-//        "OpenSans-Regular.ttf".to_string(),
-//        "Welcome to rust-pushrod!".to_string(),
-//        32,
-//    );
-//    text_widget.set_origin(14, 8);
-//    text_widget.set_size(400, 40);
-//    text_widget.set_color([0.75, 0.75, 1.0, 1.0]);
-//    text_widget.set_text_color([0.75, 0.25, 1.0, 1.0]);
-//    prod.widget_store.add_widget(Box::new(text_widget));
+    //    let mut text_widget = TextWidget::new(
+    //        factory,
+    //        "OpenSans-Regular.ttf".to_string(),
+    //        "Welcome to rust-pushrod!".to_string(),
+    //        32,
+    //    );
+    //    text_widget.set_origin(14, 8);
+    //    text_widget.set_size(400, 40);
+    //    text_widget.set_color([0.75, 0.75, 1.0, 1.0]);
+    //    text_widget.set_text_color([0.75, 0.25, 1.0, 1.0]);
+    //    prod.widget_store.add_widget(Box::new(text_widget));
 
-    let mut base_widget = BaseWidget::new();
+    let mut base_widget = CanvasWidget::new();
     base_widget.set_origin(50, 80);
     base_widget.set_size(200, 200);
     base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
@@ -58,6 +58,9 @@ fn main() {
     box_widget.set_size(200, 200);
     box_widget.set_color([0.0, 1.0, 0.0, 1.0]);
     box_widget.set_border([1.0, 0.0, 0.0, 1.0], 4);
+    box_widget.on_key_pressed(Box::new(|widget_id, key, state| {
+        eprintln!("Key {:?}; State {:?}", key, state);
+    }));
     prod.widget_store.add_widget(Box::new(box_widget));
 
     let mut box_1 = BoxWidget::new();
@@ -92,10 +95,10 @@ fn main() {
     let mut timer = TimerWidget::new();
     timer.set_timeout(1000);
     timer.set_enabled(true);
-//    timer.on_timeout(Box::new(|| eprintln!("Timer.")));
+    //    timer.on_timeout(Box::new(|| eprintln!("Timer.")));
     prod.widget_store.add_widget(Box::new(timer));
 
-//    prod.add_event_listener_for_window(Box::new(ExampleListener::new()));
+    //    prod.add_event_listener_for_window(Box::new(ExampleListener::new()));
 
     // Runs the main event loop
     prod.run();