+++ src/lib.rs
@@ -15,19 +15,23 @@
 //! Pushrod is a Cross Platform UI Widget Library for Rust.
 //!
 //! It is intended to be lightweight, easy to use, and easy to understand.  Pushrod draws
-//! inspiration from 16-bit GUI-based systems and other libraries over the years.
+//! inspiration from 16-bit GUI-based systems and other GUI libraries over the years.
 //!
-//! Pushrod uses Piston as its main window drawing and event loop functionality.  It utilizes
-//! piston2d-opengl graphics so that - eventually - the graphics can be represented as 3D
-//! poly objects with textures as canvases.  It also utilizes the `gfx_core` library for
-//! graphics functionality, as well as the `gfx_device_gl` support.
-//!
+//! # Dependencies
 //! Pushrod uses the following dependencies:
 //! ```ignore
 //! [dependencies]
 //! piston_window = "^0.89.0"
 //! find_folder = "^0.3.0"
 //! ```
+//!
+//! # Core Components
+//!
+//! # Callbacks
+//!
+//! # Events
+//!
+//! # Widgets
 
 /// Main module containing the run loop for the UI components, containers for windows and
 /// `Widget` trait objects, and so on.  Contains the core elements required to build
@@ -48,4 +52,5 @@ pub mod event;
 /// - Box Widget (for drawing a plain background with a box and a colored border)
 /// - Text Widget (for drawing text)
 /// - Timer Widget (for performing timer operations)
+/// - Image Widget (for drawing images)
 pub mod widget;