+++ src/widget/widget.rs
@@ -29,7 +29,7 @@ use crate::widget::config::*;
 /// You _should_ override `draw`, but you are not required to.  (If you don't, however, your
 /// widget won't really do much.)
 ///
-/// If you want a blank base widget, refer to the `BaseWidget`, which will create a
+/// If you want a blank base widget, refer to the `CanvasWidget`, which will create a
 /// base widget that paints the contents of its bounds with whatever color has been
 /// specified with `set_color`.
 pub trait Widget {
@@ -168,22 +168,6 @@ pub trait Widget {
         }
     }
 
-    /// Indicates to the underlying drawing mechanism as to whether or not this `Widget` needs to
-    /// have drawing clipping automatically applied.
-    fn set_autoclip(&mut self, clip: bool) {
-        self.config()
-            .set(CONFIG_AUTOCLIP, WidgetConfig::Autoclip { clip });
-        self.invalidate();
-    }
-
-    /// Retrieves the auto clip flag.
-    fn get_autoclip(&mut self) -> bool {
-        match self.config().get(CONFIG_AUTOCLIP) {
-            Some(WidgetConfig::Autoclip { ref clip }) => clip.clone(),
-            _ => false,
-        }
-    }
-
     // Callbacks
 
     /// Performs a callback stored in the `CallbackStore` for this `Widget`, but only for the
@@ -206,7 +190,28 @@ pub trait Widget {
         }
     }
 
+    fn perform_key_callback(
+        &mut self,
+        callback_id: u32,
+        widget_id: i32,
+        key: Key,
+        state: ButtonState,
+    ) {
+        match self.callbacks().get(callback_id) {
+            CallbackTypes::KeyCallback { callback } => {
+                callback(widget_id, key.clone(), state.clone())
+            }
+            _ => (),
+        }
+    }
+
     // Callback Triggers
+
+    /// Called when a keyboard event takes place within the bounds of a widget.  Includes the widget
+    /// ID, the key code that was affected, and its state - pressed or released.
+    fn key_pressed(&mut self, widget_id: i32, key: &Key, state: &ButtonState) {
+        self.perform_key_callback(CALLBACK_KEY_PRESSED, widget_id, key.clone(), state.clone());
+    }
 
     /// Called when a mouse enters the bounds of the widget.  Includes the widget ID.  Only override
     /// if you want to signal a mouse enter event.
@@ -233,6 +238,12 @@ pub trait Widget {
     }
 
     // Callback Setters
+    fn on_key_pressed(&mut self, callback: KeyCallback) {
+        self.callbacks().put(
+            CALLBACK_KEY_PRESSED,
+            CallbackTypes::KeyCallback { callback },
+        );
+    }
 
     /// Sets the closure action to be performed when a mouse enters a `Widget`.
     fn on_mouse_entered(&mut self, callback: SingleCallback) {
@@ -279,12 +290,7 @@ pub trait Widget {
 
         rectangle(
             self.get_color(),
-            [
-                origin.x as f64,
-                origin.y as f64,
-                size.w as f64,
-                size.h as f64,
-            ],
+            [0.0 as f64, 0.0 as f64, size.w as f64, size.h as f64],
             c.transform,
             g,
         );
@@ -293,16 +299,16 @@ pub trait Widget {
     }
 }
 
-/// This is the `BaseWidget`, which contains a top-level widget for display.  It does
+/// This is the `CanvasWidget`, which contains a top-level widget for display.  It does
 /// not contain any special logic other than being a base for a display layer.
-pub struct BaseWidget {
+pub struct CanvasWidget {
     config: Configurable,
     callbacks: CallbackStore,
 }
 
-/// Implementation of the constructor for the `PushrodBaseWidget`.  Creates a new base widget
+/// Implementation of the constructor for the `CanvasWidget`.  Creates a new base widget
 /// that can be positioned anywhere on the screen.
-impl BaseWidget {
+impl CanvasWidget {
     pub fn new() -> Self {
         Self {
             config: Configurable::new(),
@@ -311,7 +317,7 @@ impl BaseWidget {
     }
 }
 
-/// Implementation of the `BaseWidget` object with the `Widget` traits implemented.
+/// Implementation of the `CanvasWidget` object with the `Widget` traits implemented.
 /// This function only implements `config` and `callbacks`, which are used as a base for
 /// all `Widget`s.
 ///
@@ -328,7 +334,7 @@ impl BaseWidget {
 /// #           .build()
 /// #           .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error)));
 /// #
-///    let mut base_widget = BaseWidget::new();
+///    let mut base_widget = CanvasWidget::new();
 ///
 ///    base_widget.set_origin(100, 100);
 ///    base_widget.set_size(200, 200);
@@ -339,7 +345,7 @@ impl BaseWidget {
 ///
 ///    eprintln!("Added widget: ID={}", widget_id);
 ///
-///    let mut base_widget_2 = BaseWidget::new();
+///    let mut base_widget_2 = CanvasWidget::new();
 ///
 ///    base_widget_2.set_origin(125, 125);
 ///    base_widget_2.set_size(100, 100);
@@ -349,7 +355,7 @@ impl BaseWidget {
 ///    let widget_id_2 = prod.widget_store.add_widget_to_parent(Box::new(base_widget_2), widget_id);
 /// # }
 /// ```
-impl Widget for BaseWidget {
+impl Widget for CanvasWidget {
     fn config(&mut self) -> &mut Configurable {
         &mut self.config
     }