+++ RELEASE_NOTES.md
@@ -1,5 +1,16 @@
 # Pushrod Releases
 
+## 0.1.16
+
+- Keyboard events ticket (#43) completed by dannyfritz, merged to master.
+- Modified code so that origin of drawing is 0x0 relative to the widget (#22)
+- Changed Cargo.toml to only use major version of Piston, minor releases are automatically used.
+- Updated README.
+- Moved the context origin to 0x0 outside of the draw loop (#67)
+- Renamed BaseWidget to CanvasWidget (#27)
+- Removed autoclip config, as this will be automatic. (#20)
+- Context reset automatically takes place before each draw (#21)
+
 ## 0.1.15
 
 - Made a new `BlankCallback` type for timer